--- source-hello-world/CHANGELOG.md
@@ -1,4 +1,0 @@
-# source-hello-world
-
-## v1, 2022-07-27
-- Beginning of changelog.
