+++ source-hello-world/main.go
@@ -1,148 +1,197 @@
 package main
 
 import (
+	"context"
 	"encoding/json"
 	"fmt"
+	"strconv"
+	"strings"
 	"time"
 
-	"github.com/estuary/flow/go/protocols/airbyte"
+	schemagen "github.com/estuary/connectors/go/schema-gen"
+	boilerplate "github.com/estuary/connectors/source-boilerplate"
+	pc "github.com/estuary/flow/go/protocols/capture"
+	pf "github.com/estuary/flow/go/protocols/flow"
+	log "github.com/sirupsen/logrus"
 )
 
-type Config struct {
-	Greetings int `json:"greetings"`
-}
-
-type State struct {
-	Cursor int `json:"cursor"`
-}
-
-func (c *Config) Validate() error {
-	return nil
-}
-
-func (c *State) Validate() error {
-	return nil
-}
-
-const configSchema = `{
-	"$schema": "http://json-schema.org/draft-07/schema#",
-	"title":   "An Infinite Stream of Hello Worlds",
-	"type":    "object",
-	"properties": {
-		"greetings": {
-			"type": ["integer", "null"],
-			"default": null
-		}
-	}
-}`
-
-const greetingSchema = `{
-	"type":"object",
-	"properties": {
-		"count":   {"type": "integer"},
-		"message": {"type": "string"}
-	},
-	"required": ["count", "message"]
-}`
-
 func main() {
-	airbyte.RunMain(spec, doCheck, doDiscover, doRead)
+	boilerplate.RunMain(new(driver))
 }
 
-var spec = airbyte.Spec{
-	SupportsIncremental:           true,
-	SupportedDestinationSyncModes: airbyte.AllDestinationSyncModes,
-	ConnectionSpecification:       json.RawMessage(configSchema),
-	DocumentationURL:              "https://go.estuary.dev/source-hello-world",
+type config struct {
+	Rate float32 `json:"rate" jsonschema:"title=Message Rate,description=Message rate in documents per second,default=1"`
 }
 
-func doCheck(args airbyte.CheckCmd) error {
-	var result = &airbyte.ConnectionStatus{
-		Status: airbyte.StatusSucceeded,
+func (c config) Validate() error {
+	if c.Rate <= 0 {
+		return fmt.Errorf("message rate must be positive (got %f)", c.Rate)
+	}
+	return nil
+}
+
+type resource struct {
+	Name   string `json:"name" jsonschema:"title=Stream Name,description=Stream name to include in generated documents,default=greetings"`
+	Prefix string `json:"prefix" jsonschema:"title=Message Prefix,description=Constant prefix to add to every message,default=Hello {}"`
+}
+
+func (r resource) Validate() error {
+	return nil
+}
+
+type message struct {
+	Timestamp time.Time `json:"ts" jsonschema:"title=Timestamp,description=The time at which this message was generated"`
+	Message   string    `json:"message" jsonschema:"title=Message,description=A human-readable message"`
+}
+
+type state struct {
+	Cursor int `json:"cursor"` // The last message generated
+}
+
+type capture struct {
+	Stream   *boilerplate.PullOutput
+	Bindings []resource
+	Config   config
+	State    state
+}
+
+type driver struct{}
+
+var _ boilerplate.Connector = &driver{}
+
+func (driver) Spec(ctx context.Context, req *pc.Request_Spec) (*pc.Response_Spec, error) {
+	var endpointSchema, err = schemagen.GenerateSchema("Hello World", &config{}).MarshalJSON()
+	if err != nil {
+		return nil, fmt.Errorf("generating endpoint schema: %w", err)
 	}
 
-	if err := args.ConfigFile.Parse(new(Config)); err != nil {
-		result.Status = airbyte.StatusFailed
-		result.Message = err.Error()
+	resourceSchema, err := schemagen.GenerateSchema("Example Resource Spec", &resource{}).MarshalJSON()
+	if err != nil {
+		return nil, fmt.Errorf("generating resource schema: %w", err)
 	}
 
-	return airbyte.NewStdoutEncoder().Encode(airbyte.Message{
-		Type:             airbyte.MessageTypeConnectionStatus,
-		ConnectionStatus: result,
-	})
+	return &pc.Response_Spec{
+		ConfigSchemaJson:         json.RawMessage(endpointSchema),
+		ResourceConfigSchemaJson: json.RawMessage(resourceSchema),
+		DocumentationUrl:         "https://go.estuary.dev/source-hello-world",
+	}, nil
 }
 
-func doDiscover(args airbyte.DiscoverCmd) error {
-	if err := args.ConfigFile.Parse(new(Config)); err != nil {
-		return err
+func (driver) Validate(ctx context.Context, req *pc.Request_Validate) (*pc.Response_Validated, error) {
+	var cfg config
+	if err := pf.UnmarshalStrict(req.ConfigJson, &cfg); err != nil {
+		return nil, fmt.Errorf("parsing endpoint config: %w", err)
 	}
 
-	var catalog = new(airbyte.Catalog)
-	catalog.Streams = append(catalog.Streams, airbyte.Stream{
-		Name:                    "greetings",
-		JSONSchema:              json.RawMessage(greetingSchema),
-		SupportedSyncModes:      []airbyte.SyncMode{airbyte.SyncModeIncremental},
-		SourceDefinedCursor:     true,
-		SourceDefinedPrimaryKey: [][]string{{"count"}},
-	})
-
-	var encoder = airbyte.NewStdoutEncoder()
-	return encoder.Encode(airbyte.Message{
-		Type:    airbyte.MessageTypeCatalog,
-		Catalog: catalog,
-	})
-}
-
-func doRead(args airbyte.ReadCmd) error {
-	var config Config
-	var state State
-	var catalog airbyte.ConfiguredCatalog
-
-	if err := args.ConfigFile.Parse(&config); err != nil {
-		return err
-	} else if err := args.CatalogFile.Parse(&catalog); err != nil {
-		return err
-	} else if args.StateFile != "" {
-		if err := args.StateFile.Parse(&state); err != nil {
-			return err
+	var out []*pc.Response_Validated_Binding
+	for _, binding := range req.Bindings {
+		var res resource
+		if err := pf.UnmarshalStrict(binding.ResourceConfigJson, &res); err != nil {
+			return nil, fmt.Errorf("parsing resource config: %w", err)
 		}
-	}
 
-	var enc = airbyte.NewStdoutEncoder()
-	var now = time.Now()
-	for {
-		var b, err = json.Marshal(struct {
-			Count   int    `json:"count"`
-			Message string `json:"message"`
-		}{
-			state.Cursor,
-			fmt.Sprintf("Hello #%d", state.Cursor),
+		out = append(out, &pc.Response_Validated_Binding{
+			ResourcePath: []string{res.Name},
 		})
+	}
+	return &pc.Response_Validated{Bindings: out}, nil
+}
+
+func (driver) Discover(ctx context.Context, req *pc.Request_Discover) (*pc.Response_Discovered, error) {
+	messageSchema, err := schemagen.GenerateSchema("Example Output Record", &message{}).MarshalJSON()
+	if err != nil {
+		return nil, fmt.Errorf("generating message schema: %w", err)
+	}
+
+	resourceJSON, err := json.Marshal(resource{Name: "greetings", Prefix: "Hello {}!"})
+	if err != nil {
+		return nil, fmt.Errorf("serializing resource json: %w", err)
+	}
+
+	return &pc.Response_Discovered{
+		Bindings: []*pc.Response_Discovered_Binding{{
+			RecommendedName:    "events",
+			ResourceConfigJson: resourceJSON,
+			DocumentSchemaJson: messageSchema,
+			Key:                []string{"/ts"},
+		}},
+	}, nil
+}
+
+func (d driver) Apply(ctx context.Context, req *pc.Request_Apply) (*pc.Response_Applied, error) {
+	return &pc.Response_Applied{
+		ActionDescription: "this is a dummy capture so nothing happened",
+	}, nil
+}
+
+func (driver) Pull(open *pc.Request_Open, stream *boilerplate.PullOutput) error {
+	log.Debug("connector started")
+
+	var cfg config
+	if err := pf.UnmarshalStrict(open.Capture.ConfigJson, &cfg); err != nil {
+		return fmt.Errorf("parsing endpoint config: %w", err)
+	}
+
+	var checkpoint state
+	if open.StateJson != nil {
+		if err := json.Unmarshal(open.StateJson, &checkpoint); err != nil {
+			return fmt.Errorf("parsing driver checkpoint: %w", err)
+		}
+	}
+
+	var resources []resource
+	for _, binding := range open.Capture.Bindings {
+		var res resource
+		if err := pf.UnmarshalStrict(binding.ResourceConfigJson, &res); err != nil {
+			return fmt.Errorf("parsing resource config: %w", err)
+		}
+		resources = append(resources, res)
+	}
+
+	var capture = &capture{
+		Stream:   stream,
+		Bindings: resources,
+		Config:   cfg,
+		State:    checkpoint,
+	}
+	return capture.Run()
+}
+
+func (c *capture) Run() error {
+	if err := c.Stream.Ready(false); err != nil {
+		return err
+	}
+
+	var interval = time.Duration(float64(time.Second) / float64(c.Config.Rate))
+	for {
+		select {
+		case <-c.Stream.Context().Done():
+			log.Info("shutting down due to context cancellation")
+			return nil
+		default:
+		}
+
+		for idx, binding := range c.Bindings {
+			var text = strings.Replace(binding.Prefix, "{}", strconv.Itoa(c.State.Cursor), -1)
+			var doc = &message{
+				Timestamp: time.Now(),
+				Message:   text,
+			}
+			bs, err := json.Marshal(doc)
+			if err != nil {
+				return fmt.Errorf("error serializing document: %w", err)
+			} else if err := c.Stream.Documents(idx, json.RawMessage(bs)); err != nil {
+				return err
+			}
+		}
+		c.State.Cursor++
+
+		checkpointJsonBytes, err := json.Marshal(c.State)
 		if err != nil {
+			return fmt.Errorf("error serializing state: %w", err)
+		} else if err := c.Stream.Checkpoint(checkpointJsonBytes, true); err != nil {
 			return err
 		}
-		state.Cursor++
-
-		if err = enc.Encode(&airbyte.Message{
-			Type: airbyte.MessageTypeRecord,
-			Record: &airbyte.Record{
-				Stream:    "greetings",
-				EmittedAt: now.UTC().UnixNano() / int64(time.Millisecond),
-				Data:      b,
-			},
-		}); err != nil {
-			return err
-		}
-
-		if b, err = json.Marshal(state); err != nil {
-			return err
-		} else if err = enc.Encode(airbyte.Message{
-			Type:  airbyte.MessageTypeState,
-			State: &airbyte.State{Data: b},
-		}); err != nil {
-			return err
-		}
-
-		now = <-time.After(time.Millisecond * 500)
+		time.Sleep(interval)
 	}
 }
