--- source-nativehello/Dockerfile
@@ -1,41 +1,0 @@
-# Build Stage
-################################################################################
-FROM golang:1.20-buster as builder
-
-WORKDIR /builder
-
-# Download & compile dependencies early. Doing this separately allows for layer
-# caching opportunities when no dependencies are updated.
-COPY go.* ./
-RUN go mod download
-
-COPY go                 ./go
-COPY source-boilerplate ./source-boilerplate
-COPY source-nativehello ./source-nativehello
-
-# Run the unit tests.
-RUN go test -v ./source-nativehello/...
-
-# Build the connector.
-RUN go build -o ./connector -v ./source-nativehello
-
-
-# Runtime Stage
-################################################################################
-FROM ghcr.io/estuary/base-image:v1
-
-WORKDIR /connector
-ENV PATH="/connector:$PATH"
-
-COPY --from=busybox:latest /bin/sh /bin/sh
-
-# Bring in the compiled connector artifact from the builder.
-COPY --from=builder /builder/connector ./source-nativehello
-
-LABEL FLOW_RUNTIME_PROTOCOL=capture
-LABEL CONNECTOR_PROTOCOL=flow-capture
-
-# Avoid running the connector as root.
-USER nonroot:nonroot
-
-ENTRYPOINT ["/connector/source-nativehello"]
